+++ src/features/components/molecules/protected-route/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import PageNotFound from 'features/pages/404-error'
+import PageNotFound from 'features/pages/404'
 import InitialLoader from 'components/elements/dot-loader'
 import Layout from 'features/components/templates/layout'
