+++ src/features/pages/404/content/types.ts
