+++ src/pages/deriv-go/index.tsx
@@ -4,7 +4,7 @@ import Banner from './_banner'
 import OtherApps from './_other-apps'
 import WhatIsDerivGo from './_what-is-deriv-go'
 import WhyTradeDerivGo from './_why-trade-deriv-go'
-import PageNotFound from 'features/pages/404-error'
+import PageNotFound from 'features/pages/404'
 import { SEO } from 'components/containers'
 import Roadmap, { TPortal } from 'components/elements/roadmap'
 import useRegion from 'components/hooks/use-region'
