+++ src/pages/derivx/index.tsx
@@ -6,7 +6,7 @@ import WhyTradeDerivX from './_why-trade-derivx'
 import StartDerivX from './_start-derivx'
 import DerivXGetApp from './_get-derivx'
 import OurPlatforms from './_other-apps'
-import PageNotFound from 'features/pages/404-error'
+import PageNotFound from 'features/pages/404'
 import { SEO } from 'components/containers'
 import Layout from 'components/layout/layout'
 import { WithIntl } from 'components/localization'
