+++ src/pages/p2p/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Loadable from '@loadable/component'
 import Hero from './components/_hero'
 import DP2P, { P2PType } from './components/_dp2p'
-import PageNotFound from 'features/pages/404-error'
+import PageNotFound from 'features/pages/404'
 import InitialLoader from 'components/elements/dot-loader'
 import Roadmap, { TPortal } from 'components/elements/roadmap'
 import { usePageLoaded } from 'components/hooks/use-page-loaded'
