+++ src/pages/partners/payment-agent/index.tsx
@@ -5,7 +5,7 @@ import Hero from './_pa-hero'
 import TapInto from './_tap-into'
 import { faq_schema } from './_faq-schema'
 import PremiumPaymentAgent from './_premium_payment_agent'
-import PageNotFound from 'features/pages/404-error'
+import PageNotFound from 'features/pages/404'
 import Layout from 'components/layout/layout'
 import { SEO, MetaAttributesType } from 'components/containers'
 import { Divider } from 'components/elements'
