+++ src/pages/trade-types/accumulators/index.tsx
@@ -3,7 +3,7 @@ import Loadable from '@loadable/component'
 import styled from 'styled-components'
 import { Hero } from '../components/_style'
 import WhatAreAccumulators from './_what-are-accumulators'
-import PageNotFound from 'features/pages/404-error'
+import PageNotFound from 'features/pages/404'
 import { SEO, SmallContainer, MetaAttributesType } from 'components/containers'
 import Layout from 'components/layout/layout'
 import Button from 'components/custom/_button'
